+++ src/python/pants/backend/go/build.py
@@ -10,7 +10,6 @@ from typing import List, Optional
 
 from pants.backend.go.pkg import (
     ResolvedGoPackage,
-    ResolveExternalGoPackageRequest,
     ResolveGoPackageRequest,
     is_first_party_package_target,
     is_third_party_package_target,
@@ -23,6 +22,7 @@ from pants.backend.go.target_types import (
 from pants.backend.go.util_rules.external_module import (
     DownloadedExternalModule,
     DownloadExternalModuleRequest,
+    ResolveExternalGoPackageRequest,
 )
 from pants.backend.go.util_rules.import_analysis import ResolvedImportPathsForGoLangDistribution
 from pants.backend.go.util_rules.sdk import GoSdkProcess
