+++ src/python/pants/backend/go/pkg.py
@@ -6,20 +6,11 @@ import json
 import logging
 from dataclasses import dataclass
 
-import ijson
-
 from pants.backend.go.target_types import (
-    GoExternalModulePathField,
-    GoExternalModuleVersionField,
     GoExternalPackageDependencies,
-    GoExternalPackageImportPathField,
     GoImportPath,
     GoModuleSources,
     GoPackageSources,
-)
-from pants.backend.go.util_rules.external_module import (
-    DownloadedExternalModule,
-    DownloadExternalModuleRequest,
 )
 from pants.backend.go.util_rules.go_mod import (
     FindNearestGoModuleRequest,
@@ -30,23 +21,10 @@ from pants.backend.go.util_rules.go_mod import (
 from pants.backend.go.util_rules.sdk import GoSdkProcess
 from pants.build_graph.address import Address
 from pants.core.util_rules.source_files import SourceFiles, SourceFilesRequest
-from pants.engine.fs import (
-    AddPrefix,
-    CreateDigest,
-    Digest,
-    DigestContents,
-    DigestSubset,
-    FileContent,
-    GlobExpansionConjunction,
-    GlobMatchErrorBehavior,
-    MergeDigests,
-    PathGlobs,
-)
 from pants.engine.internals.selectors import Get, MultiGet
 from pants.engine.process import ProcessResult
 from pants.engine.rules import collect_rules, rule
 from pants.engine.target import Target, WrappedTarget
-from pants.util.ordered_set import FrozenOrderedSet, OrderedSet
 
 logger = logging.getLogger(__name__)
 
@@ -192,11 +170,6 @@ class ResolveGoPackageRequest:
     address: Address
 
 
-@dataclass(frozen=True)
-class ResolveExternalGoPackageRequest:
-    address: Address
-
-
 def error_to_string(d: dict) -> str:
     pos = d.get("Pos", "")
     if pos:
@@ -281,144 +254,6 @@ async def resolve_go_package(
         address=request.address,
         module_address=owning_go_module_result.module_address,
     )
-
-
-@rule
-async def resolve_external_go_package(
-    request: ResolveExternalGoPackageRequest,
-) -> ResolvedGoPackage:
-    wrapped_target = await Get(WrappedTarget, Address, request.address)
-    target = wrapped_target.target
-
-    import_path = target[GoExternalPackageImportPathField].value
-    module_path = target[GoExternalModulePathField].value
-    module_version = target[GoExternalModuleVersionField].value
-
-    module = await Get(
-        DownloadedExternalModule,
-        DownloadExternalModuleRequest(
-            path=module_path,
-            version=module_version,
-        ),
-    )
-
-    assert import_path.startswith(module_path)
-    subpath = import_path[len(module_path) :]
-
-    result = await Get(
-        ProcessResult,
-        GoSdkProcess(
-            input_digest=module.digest,
-            command=("list", "-json", f"./{subpath}"),
-            description="Resolve _go_external_package metadata.",
-        ),
-    )
-
-    metadata = json.loads(result.stdout)
-    return ResolvedGoPackage.from_metadata(
-        metadata,
-        import_path=import_path,
-        address=request.address,
-        module_address=None,
-        module_path=module_path,
-        module_version=module_version,
-    )
-
-
-@dataclass(frozen=True)
-class ResolveExternalGoModuleToPackagesRequest:
-    path: str
-    version: str
-    go_sum_digest: Digest
-
-
-@dataclass(frozen=True)
-class ResolveExternalGoModuleToPackagesResult:
-    # TODO: Consider using DeduplicatedCollection if this is the only field.
-    packages: FrozenOrderedSet[ResolvedGoPackage]
-
-
-@rule
-async def resolve_external_module_to_go_packages(
-    request: ResolveExternalGoModuleToPackagesRequest,
-) -> ResolveExternalGoModuleToPackagesResult:
-    module_path = request.path
-    assert module_path
-    module_version = request.version
-    assert module_version
-
-    downloaded_module = await Get(
-        DownloadedExternalModule,
-        DownloadExternalModuleRequest(path=module_path, version=module_version),
-    )
-    sources_digest = await Get(Digest, AddPrefix(downloaded_module.digest, "__sources__"))
-
-    # TODO: Super hacky merge of go.sum from both digests. We should really just pass in the fully-resolved
-    # go.sum and use that, but this allows the go.sum from the downloaded module to have some effect. Not sure
-    # if that is right call, but hackity hack!
-    left_digest_contents = await Get(DigestContents, Digest, sources_digest)
-    left_go_sum_contents = b""
-    for fc in left_digest_contents:
-        if fc.path == "__sources__/go.sum":
-            left_go_sum_contents = fc.content
-            break
-
-    go_sum_only_digest = await Get(
-        Digest, DigestSubset(request.go_sum_digest, PathGlobs(["go.sum"]))
-    )
-    go_sum_prefixed_digest = await Get(Digest, AddPrefix(go_sum_only_digest, "__sources__"))
-    right_digest_contents = await Get(DigestContents, Digest, go_sum_prefixed_digest)
-    right_go_sum_contents = b""
-    for fc in right_digest_contents:
-        if fc.path == "__sources__/go.sum":
-            right_go_sum_contents = fc.content
-            break
-    go_sum_contents = left_go_sum_contents + b"\n" + right_go_sum_contents
-    go_sum_digest = await Get(
-        Digest,
-        CreateDigest(
-            [
-                FileContent(
-                    path="__sources__/go.sum",
-                    content=go_sum_contents,
-                )
-            ]
-        ),
-    )
-
-    sources_digest_no_go_sum = await Get(
-        Digest,
-        DigestSubset(
-            sources_digest,
-            PathGlobs(
-                ["!__sources__/go.sum", "__sources__/**"],
-                conjunction=GlobExpansionConjunction.all_match,
-                glob_match_error_behavior=GlobMatchErrorBehavior.error,
-                description_of_origin="FUNKY",
-            ),
-        ),
-    )
-
-    input_digest = await Get(Digest, MergeDigests([sources_digest_no_go_sum, go_sum_digest]))
-
-    result = await Get(
-        ProcessResult,
-        GoSdkProcess(
-            input_digest=input_digest,
-            command=("list", "-json", "./..."),
-            working_dir="__sources__",
-            description=f"Resolve packages in Go external module {module_path}@{module_version}",
-        ),
-    )
-
-    packages: OrderedSet[ResolvedGoPackage] = OrderedSet()
-    for metadata in ijson.items(result.stdout, "", multiple_values=True):
-        package = ResolvedGoPackage.from_metadata(
-            metadata, module_path=module_path, module_version=module_version
-        )
-        packages.add(package)
-
-    return ResolveExternalGoModuleToPackagesResult(packages=FrozenOrderedSet(packages))
 
 
 def rules():
