+++ src/python/pants/backend/go/target_type_rules.py
@@ -7,11 +7,7 @@ from collections import defaultdict
 from dataclasses import dataclass
 
 from pants.backend.go import pkg
-from pants.backend.go.pkg import (
-    ResolvedGoPackage,
-    ResolveExternalGoPackageRequest,
-    ResolveGoPackageRequest,
-)
+from pants.backend.go.pkg import ResolvedGoPackage, ResolveGoPackageRequest
 from pants.backend.go.target_types import (
     GoExternalPackageDependencies,
     GoImportPath,
@@ -19,6 +15,7 @@ from pants.backend.go.target_types import (
     GoPackageSources,
 )
 from pants.backend.go.util_rules import import_analysis
+from pants.backend.go.util_rules.external_module import ResolveExternalGoPackageRequest
 from pants.backend.go.util_rules.go_mod import FindNearestGoModuleRequest, ResolvedOwningGoModule
 from pants.backend.go.util_rules.import_analysis import ResolvedImportPathsForGoLangDistribution
 from pants.base.specs import (
